Move grammars and parsers to dedicated folders
Move parsing related tests under the new folder.
