+++ .gitignore
@@ -135,7 +135,7 @@ docs/
 
 # Generated grammars
 
-src/sonic-weave-ast.js
-src/sonic-weave-chord.js
+src/parser/sonic-weave-ast.js
+src/parser/sonic-weave-chord.js
 src/scale-workshop-2-ast.js
-src/paren-counter.js
+src/parser/paren-counter.js
