+++ package.json
@@ -36,10 +36,10 @@
   "scripts": {
     "lint": "gts lint",
     "clean": "gts clean",
-    "compile-sw2-parser": "peggy src/scale-workshop-2.pegjs -o src/scale-workshop-2-ast.js",
-    "compile-chord-parser": "peggy src/sonic-weave-chord.pegjs src/base.pegjs -o src/sonic-weave-chord.js",
-    "compile-sonic-weave-parser": "peggy src/sonic-weave.pegjs src/base.pegjs -o src/sonic-weave-ast.js",
-    "compile-paren-counter": "peggy src/paren-counter.pegjs src/base.pegjs -o src/paren-counter.js",
+    "compile-sw2-parser": "peggy src/grammars/scale-workshop-2.pegjs -o src/scale-workshop-2-ast.js",
+    "compile-chord-parser": "peggy src/grammars/sonic-weave-chord.pegjs src/grammars/base.pegjs -o src/parser/sonic-weave-chord.js",
+    "compile-sonic-weave-parser": "peggy src/grammars/sonic-weave.pegjs src/grammars/base.pegjs -o src/parser/sonic-weave-ast.js",
+    "compile-paren-counter": "peggy src/grammars/paren-counter.pegjs src/grammars/base.pegjs -o src/parser/paren-counter.js",
     "compile-parsers": "npm run compile-sonic-weave-parser && npm run compile-chord-parser && npm run compile-sw2-parser && npm run compile-paren-counter",
     "precompile": "npm run compile-parsers",
     "compile": "tsc",
