+++ src/grammars/paren-counter.pegjs
