+++ src/parser/__tests__/paren-counter.spec.ts
